GRP / build engine rendering fixes and changes

+++ Engine/Models/TextureOptions.cs
@@ -6,8 +6,8 @@
         public int XOffset { get; set; }
         public int YOffset { get; set; }
         public float Alpha { get; set; }
-        public int? XScale { get; set; }
-        public int? YScale { get; set; }
+        public float? XScale { get; set; }
+        public float? YScale { get; set; }
         public TextureRenderingOptions RenderingOptions { get; set; }
     }
